+++ src/System/Taffybar/Context.hs
@@ -88,7 +88,7 @@ data TaffybarConfig = TaffybarConfig
   { dbusClientParam :: Maybe DBus.Client
   , startupHook :: TaffyIO ()
   , getBarConfigsParam :: BarConfigGetter
-  , cssPath :: Maybe FilePath
+  , cssPaths :: [FilePath]
   , errorMsg :: Maybe String
   }
 
@@ -101,7 +101,7 @@ defaultTaffybarConfig = TaffybarConfig
   { dbusClientParam = Nothing
   , startupHook = return ()
   , getBarConfigsParam = return []
-  , cssPath = Nothing
+  , cssPaths = []
   , errorMsg = Nothing
   }